requirements.txt
streamlit
opencv-python-headless
pytesseract
pillow
numpy